refactor: separate top bar composable
