+++ app/src/main/java/com/android/wildex/ui/report/ReportScreen.kt
@@ -22,7 +22,6 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.LocationOn
-import androidx.compose.material.icons.outlined.Notifications
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -32,7 +31,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.MaterialTheme.colorScheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.pulltorefresh.PullToRefreshBox
 import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.runtime.Composable
@@ -47,7 +45,6 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
@@ -164,58 +161,6 @@ fun ReportScreen(
           }
         }
       }
-}
-
-/**
- * A composable that displays the top bar of the ReportScreen.
- *
- * @param userId The ID of the user.
- * @param userType The type of the user.
- * @param userProfilePictureURL The URL of the user's profile picture.
- * @param onProfileClick The function to be called when the profile is clicked.
- * @param onNotificationClick The function to be called when the notification button is clicked.
- */
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ReportScreenTopBar(
-    userId: Id = "",
-    userType: UserType = UserType.REGULAR,
-    userProfilePictureURL: URL = "",
-    onProfileClick: (Id) -> Unit = {},
-    onNotificationClick: () -> Unit
-) {
-  TopAppBar(
-      title = {
-        Text(
-            modifier = Modifier.fillMaxWidth().testTag(ReportScreenTestTags.SCREEN_TITLE),
-            text =
-                when (userType) {
-                  UserType.REGULAR -> LocalContext.current.getString(R.string.report_title_regular)
-                  UserType.PROFESSIONAL ->
-                      LocalContext.current.getString(R.string.report_title_professional)
-                },
-            textAlign = TextAlign.Center,
-            color = colorScheme.primary,
-            style = MaterialTheme.typography.titleLarge)
-      },
-      navigationIcon = {
-        IconButton(
-            modifier = Modifier.testTag(ReportScreenTestTags.NOTIFICATION_BUTTON),
-            onClick = onNotificationClick) {
-              Icon(
-                  imageVector = Icons.Outlined.Notifications,
-                  contentDescription = "Notifications",
-                  tint = colorScheme.tertiary,
-                  modifier = Modifier.size(30.dp))
-            }
-      },
-      actions = {
-        ClickableProfilePicture(
-            profileId = userId,
-            profilePictureURL = userProfilePictureURL,
-            role = "user",
-            onProfile = onProfileClick)
-      })
 }
 
 /**
